refactor directory and unused class
